refactor: 각 도메인에 맞는 서비스 로직 이동 및 작성 후 재사용
