+++ src/modules/problem/problem.module.ts
@@ -4,12 +4,11 @@ import { ProblemController } from './problem.controller';
 import { ProblemService } from './problem.service';
 import { Problem } from '@/modules/problem/entities';
 import { TestCase } from '@/modules/problem/entities';
-import { SubmitCode } from '@/modules/submit/entities';
-import { User } from '@/modules/user/entities';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([Problem, SubmitCode, User, TestCase])],
+  imports: [TypeOrmModule.forFeature([Problem, TestCase])],
   controllers: [ProblemController],
   providers: [ProblemService],
+  exports: [ProblemService, TypeOrmModule],
 })
 export class ProblemModule {}
