+++ src/modules/problem/swagger/index.ts
@@ -1,4 +1,2 @@
 export * from './ApiProblem.swagger';
 export * from './ApiProblemDetail.swagger';
-export * from './ApiProblemSubmit.swagger';
-export * from './ApiProblemSubmitStatus.swagger';
