+++ src/modules/submit/dto/SubmitCodeData.dto.ts
@@ -1,7 +1,7 @@
 import { IsString } from 'class-validator';
 import { ApiProperty } from '@nestjs/swagger';
 
-export class SubmittedCode {
+export class SubmitCodeData {
   @ApiProperty({
     name: 'code',
     description: '정답으로 제출할 코드',
