+++ src/modules/submit/dto/index.ts
@@ -1,2 +1,4 @@
+export * from './SubmitCodeData.dto';
 export * from './SubmitCodePaging.dto';
 export * from './SubmitCodeSearchOptions.dto';
+export * from './judgeInfo.dto';
