+++ src/modules/submit/swagger/index.ts
@@ -1,2 +1,4 @@
+export * from './ApiSubmitCode.swagger';
 export * from './ApiSubmitCodeList.swagger';
 export * from './ApiSubmitCodeListSize.swagger';
+export * from './ApiSubmitCodeStatus.swagger';
