+++ src/modules/user/user.service.ts
@@ -15,6 +15,12 @@ export class UserService {
     @InjectRepository(User) private readonly repo: Repository<User>,
   ) {}
 
+  async getUser(userId: number): Promise<User | null> {
+    const user = await this.repo.findOne({ where: { snsId: userId } });
+
+    return user;
+  }
+
   async getUserInfo(snsId: number): Promise<UserInfo> {
     const userInfo = await this.repo.findOne({
       where: {
